+++ datacrunch/__init__.py
@@ -1,1 +1,3 @@
 from datacrunch.datacrunch import DataCrunchClient
+
+from datacrunch._version import __version__