+++ docs/source/conf.py
@@ -14,7 +14,7 @@
 #
 import datetime
 from recommonmark.parser import CommonMarkParser
-from datacrunch.__version__ import VERSION
+from datacrunch import __version__
 import os
 import sys
 sys.path.insert(0, os.path.abspath('../../'))
@@ -27,9 +27,9 @@ copyright = f'{current_year}, DataCrunch.io'
 author = 'DataCrunch.io'
 
 # The short X.Y version
-version = VERSION
+version = __version__
 # The full version, including alpha/beta/rc tags
-release = VERSION
+release = __version__
 
 
 # -- General configuration ---------------------------------------------------